Extract run table pipeline executor
